Fix bugs with PF_RING+DNA capstats
Fixed a bug where the capstats command would not work when PF_RING+DNA
load balancing was being used (fixed by adding ",1" to the "-n" option
of pfdnacluster_master command).

Fixed another bug where capstats would fail if it was not preceded
by a "broctl start" because in that case it was being run with the
original interface, such as "dna0" (fixed by renaming the worker
interfaces at initialization, instead of when the "start" command
is run).