Change handling of nodes that crash during shutdown
When a node crashes during shutdown (i.e., the node was running when a
"stop" command was issued, and was terminated by the SIGTERM, but not
gracefully), the node status is now "stopped" instead of "crashed"
when the "stop" command is done, and a crash report will no longer
be generated in this case.

However, unchanged is the fact that the user is still notified by an
output message of what happened, and the stdout/stderr.log are still
archived in case they contain any useful clues.

Also, no changes to what happens when broctl sends SIGKILL to a node
that doesn't terminate after the SIGTERM.